unit test project awssdk.s3 package install

+++ Application/Interfaces/IBookRepository.cs
@@ -4,6 +4,6 @@ namespace Application.Interfaces;
 
 public interface IBookRepository : IRepository<Book>
 {
-    Task<Book> GetById(int id);
+    Task<Book> GetById(long id);
 
 }
